Tools: Remove conftest and move remaining utilities to loki.tools.util
