--- loki/tests/conftest.py
@@ -1,110 +1,0 @@
-# (C) Copyright 2018- ECMWF.
-# This software is licensed under the terms of the Apache Licence Version 2.0
-# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
-# In applying this licence, ECMWF does not waive the privileges and immunities
-# granted to it by virtue of its status as an intergovernmental organisation
-# nor does it submit to any jurisdiction.
-
-import contextlib
-import os
-import io
-
-
-__all__ = ['stdchannel_redirected', 'stdchannel_is_captured', 'graphviz_present']
-
-
-@contextlib.contextmanager
-def stdchannel_redirected(stdchannel, dest_filename):
-    """
-    A context manager to temporarily redirect stdout or stderr
-
-    e.g.:
-
-    ```
-    with stdchannel_redirected(sys.stderr, os.devnull):
-        if compiler.has_function('clock_gettime', libraries=['rt']):
-            libraries.append('rt')
-    ```
-
-    Source: https://stackoverflow.com/a/17753573
-
-    Note, that this only works when pytest is invoked with '--show-capture' (or '-s').
-    This can be checked using `stdchannel_is_captured(capsys)`.
-    Additionally, capturing of sys.stdout/sys.stderr needs to be disabled explicitly,
-    i.e., use the fixture `capsys` and wrap the above:
-
-    ```
-    with capsys.disabled():
-        with stdchannel_redirected(sys.stdout, 'stdout.log'):
-            function()
-    ```
-    """
-
-    def try_dup(fd):
-        try:
-            oldfd = os.dup(fd.fileno())
-        except io.UnsupportedOperation:
-            oldfd = None
-        return oldfd
-
-    def try_dup2(fd, fd2, fd_fileno=True):
-        try:
-            if fd_fileno:
-                os.dup2(fd.fileno(), fd2.fileno())
-            else:
-                os.dup2(fd, fd2.fileno())
-        except io.UnsupportedOperation:
-            pass
-
-    oldstdchannel, dest_file = None, None
-    try:
-        oldstdchannel = try_dup(stdchannel)
-        dest_file = open(dest_filename, 'w')
-        try_dup2(dest_file, stdchannel)
-
-        yield
-    finally:
-        if oldstdchannel is not None:
-            try_dup2(oldstdchannel, stdchannel, fd_fileno=False)
-        if dest_file is not None:
-            dest_file.close()
-
-
-def stdchannel_is_captured(capsys):
-    """
-    Utility function to verify if pytest captures stdout/stderr.
-
-    This hinders redirecting stdout/stderr for f2py/f90wrap functions.
-
-    Parameters
-    ----------
-    capsys :
-        The capsys fixture of the test.
-
-    Returns
-    -------
-    bool
-        `True` if pytest captures output, otherwise `False`.
-    """
-
-    capturemanager = capsys.request.config.pluginmanager.getplugin("capturemanager")
-    return capturemanager._global_capturing.out is not None
-
-
-def graphviz_present():
-    """
-    Test if graphviz is present and works
-    The import will work as long as the graphviz python wrapper is available,
-    but the underlying binaries may be missing.
-    """
-    try:
-        import graphviz as gviz # pylint: disable=import-outside-toplevel
-    except ImportError:
-        return False
-
-    try:
-        gviz.Graph().pipe()
-    except gviz.ExecutableNotFound:
-        return False
-
-    return True
